removes newDiningHall
