+++ PennMobile/src/main/java/com/pennapps/labs/pennmobile/MenuFragment.java
@@ -4,27 +4,16 @@ import android.graphics.Color;
 import android.os.Bundle;
 import android.support.annotation.StyleRes;
 import android.support.v4.app.Fragment;
-import android.support.v4.content.ContextCompat;
 import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.pennapps.labs.pennmobile.classes.DiningHall;
-import com.pennapps.labs.pennmobile.classes.NewDiningHall;
 
 import org.apache.commons.lang3.StringEscapeUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.text.WordUtils;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
@@ -53,11 +42,11 @@ public class MenuFragment extends Fragment {
     }
 
     public void fillDescriptions() {
-        for (NewDiningHall.Menu menu : mDiningHall.menus) {
+        for (DiningHall.Menu menu : mDiningHall.menus) {
             addDiningTextView(R.style.MealName, StringUtils.capitalize(menu.name));
-            for (NewDiningHall.DiningStation station : menu.stations) {
+            for (DiningHall.DiningStation station : menu.stations) {
                 addDiningTextView(R.style.DiningStation, StringUtils.capitalize(station.name));
-                for (NewDiningHall.FoodItem item : station.items) {
+                for (DiningHall.FoodItem item : station.items) {
                     addDiningTextView(R.style.FoodItem, StringEscapeUtils.unescapeXml(item.title));
                 }
             }
