+++ PennMobile/src/main/java/com/pennapps/labs/pennmobile/api/Labs.java
@@ -4,7 +4,7 @@ import com.pennapps.labs.pennmobile.classes.Building;
 import com.pennapps.labs.pennmobile.classes.BusRoute;
 import com.pennapps.labs.pennmobile.classes.BusStop;
 import com.pennapps.labs.pennmobile.classes.Course;
-import com.pennapps.labs.pennmobile.classes.NewDiningHall;
+import com.pennapps.labs.pennmobile.classes.DiningHall;
 import com.pennapps.labs.pennmobile.classes.Person;
 import com.pennapps.labs.pennmobile.classes.Review;
 import com.pennapps.labs.pennmobile.classes.Venue;
@@ -37,7 +37,7 @@ public interface Labs {
     Observable<List<Venue>> venues();
 
     @GET("/dining/daily_menu/{id}")
-    Observable<NewDiningHall> daily_menu(
+    Observable<DiningHall> daily_menu(
         @Path("id") int id);
 
     @GET("/transit/stops")
