+++ PennMobile/src/main/java/com/pennapps/labs/pennmobile/api/Serializer.java
@@ -11,7 +11,7 @@ import com.pennapps.labs.pennmobile.classes.Building;
 import com.pennapps.labs.pennmobile.classes.BusRoute;
 import com.pennapps.labs.pennmobile.classes.BusStop;
 import com.pennapps.labs.pennmobile.classes.Course;
-import com.pennapps.labs.pennmobile.classes.NewDiningHall;
+import com.pennapps.labs.pennmobile.classes.DiningHall;
 import com.pennapps.labs.pennmobile.classes.Venue;
 
 import java.lang.reflect.Type;
@@ -61,15 +61,15 @@ public class Serializer {
         }
     }
 
-    public static class MenuSerializer implements JsonDeserializer<NewDiningHall> {
+    public static class MenuSerializer implements JsonDeserializer<DiningHall> {
         @Override
-        public NewDiningHall deserialize(JsonElement je, Type type, JsonDeserializationContext jdc)
+        public DiningHall deserialize(JsonElement je, Type type, JsonDeserializationContext jdc)
                 throws JsonParseException {
             JsonElement content = je.getAsJsonObject().get("Document");
             content.getAsJsonObject().add("tblDayPart",
                     content.getAsJsonObject().get("tblMenu").getAsJsonObject().get("tblDayPart").getAsJsonArray());
             content.getAsJsonObject().remove("tblMenu");
-            return new Gson().fromJson(content, NewDiningHall.class);
+            return new Gson().fromJson(content, DiningHall.class);
         }
     }
