--- PennMobile/src/main/java/com/pennapps/labs/pennmobile/classes/NewDiningHall.java
@@ -1,45 +1,0 @@
-package com.pennapps.labs.pennmobile.classes;
-
-import com.google.gson.annotations.SerializedName;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-
-/**
- * Created by Adel on 12/18/14.
- * Class for Dining Halls with Retrofit
- * Should replace DiningHall class at some point
- */
-public class NewDiningHall {
-    public String name;
-    @SerializedName("tblDayPart") public List<Menu> menus = new ArrayList<>();
-
-    /**
-     * Created by Adel on 12/18/14.
-     * Class for a single menu, ie. Lunch, Dinner
-     */
-    public static class Menu {
-        @SerializedName("txtDayPartDescription") public String name;
-        @SerializedName("tblStation") public List<DiningStation> stations = new ArrayList<>();
-    }
-
-    /**
-     * Created by Adel on 12/18/14.
-     * Class for a station at a dining hall
-     */
-    public static class DiningStation {
-        @SerializedName("txtStationDescription") public String name;
-        @SerializedName("tblItem") public List<FoodItem> items = new ArrayList<>();
-    }
-
-    /**
-     * Created by Adel on 12/18/14.
-     * Class for Food items in Dining menus
-     */
-    public static class FoodItem {
-        @SerializedName("txtTitle") public String title;
-        @SerializedName("txtDescription") String description;
-    }
-}
